HeavyLightDecomposition Part only Basic Structure
This is the basic implementation of the heavy light decomposition. This portion contains only the basic Heavy Light Part. It does not cover query portion.  

Following is the testCase over which you can check
23 
0 1 2
0 2 2
0 4 3
2 3 5
4 5 5
0 6 2
6 7 1
6 9 3
7 8 5
6 10 2
10 14 3
10 11 5
14 15 2
11 12 3
11 13 8
6 16 1
16 17 4
17 18 4
16 19 2
19 20 2
20 21 4
19 22 6

Following are the result. Better have a good look over all the arrays used to get the better understanding.

    print(chainHead);
    print(chainNo);
    print(baseArray);
    print(posInBase);
    print(sc);
    print(subSize);

0 22 17 7 9 10 13 14 1 2 4 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1
0 8 9 9 10 10 0 3 3 4 5 5 5 6 7 7 0 2 2 0 0 0 1
-1 2 1 2 2 4 6 4 4 1 5 3 2 5 3 8 3 2 2 2 5 3 5
0 18 19 20 21 22 1 9 10 11 12 13 14 15 16 17 2 7 8 3 4 5 6
6 -1 3 -1 5 -1 16 8 -1 -1 11 12 -1 -1 15 -1 19 18 -1 20 21 -1 -1
23 1 2 1 2 1 17 2 1 1 6 3 1 1 2 1 7 2 1 4 2 1 1