Solved the GSS1 Problem from Spoj
Basic Idea of the problem is merging step which is ::

prefixofanode = maximum of prefix of left node and the sum of the left node, and the prefix of the right node
suffixofanode = maximum of suffix of right node and the sum of the right node and suffix of the left node
bestsum = maximum of the left best, right best and the left node suffix and the right node suffix

This is difficult to perceive for the first time. So be patient. I have also to be patient. BEST OF LUCK